Refactored get-model-class code to a util method
