+++ lib/enrichers/schema.relationships.enricher.js
@@ -1,8 +1,7 @@
-const fs = require('fs');
-const path = require('path');
-const { transformSchema } = require('../transformers/json-schema.transformer');
-const constants = require('../constants');
 const _ = require('lodash');
+const { transformSchema } = require('../transformers/json-schema.transformer');
+const getModelClass = require('../utils/get-model-class');
+const constants = require('../constants');
 
 /**
  * Creates copy of provided schema and enriches it with attributes that are derived from relationships
@@ -107,31 +106,7 @@ function _resolveModelProperties(
   isIncludeParentRelationships,
   modelPaths
 ) {
-  let model;
-
-  if ( _.isString(modelClass)) {
-    if (modelPaths && modelPaths.length) {
-      for (const modelPath of modelPaths) {
-        let fullPath = path.join(process.cwd(), modelPath, modelClass);
-
-        if (!fullPath.endsWith('.ts') && !fullPath.endsWith('.js'))
-          fullPath += '.js';
-
-        if (fs.existsSync(fullPath)) {
-          model = require(fullPath);
-
-          break;
-        }
-      }
-
-      if (!model)
-        throw new Error(`Failed to load modelClass ${modelClass} with modelPaths ${JSON.stringify(modelPaths)}`);
-    } else {
-      model = require(modelClass);
-    }
-  } else {
-    model = modelClass;
-  }
+  const model = getModelClass(modelClass, modelPaths);
 
   if (isIncludeRelationships && model.relationMappings) {
     return enrichSchemaWithRelationships(
