+++ lib/transformers/model.schema.transformer.js
@@ -1,7 +1,6 @@
-const fs = require('fs');
-const path = require('path');
-const constants = require('../constants');
 const _ = require('lodash');
+const getModelClass = require('../utils/get-model-class');
+const constants = require('../constants');
 
 const jsonSchemaTransformer = require('./json-schema.transformer');
 
@@ -49,33 +48,7 @@ function _fillRelationshipFields(model, processedSchema, opts) {
 }
 
 function _resolveSchemaForRelation(processedClass, relationEntry, opts) {
-  const modelPaths = processedClass.modelPaths;
-  let relatedClass;
-
-  if (_.isString(relationEntry.modelClass)) {
-    if (modelPaths && modelPaths.length) {
-      for (const modelPath of modelPaths) {
-        let fullPath = path.join(process.cwd(), modelPath, relationEntry.modelClass);
-
-        if (!fullPath.endsWith('.ts') && !fullPath.endsWith('.js'))
-          fullPath += '.js';
-
-        if (fs.existsSync(fullPath)) {
-          relatedClass = require(fullPath);
-
-          break;
-        }
-      }
-
-      if (!relatedClass)
-        throw new Error(`Failed to load modelClass ${relationEntry.modelClass} with modelPaths ${JSON.stringify(modelPaths)}`);
-    } else {
-      relatedClass = require(relationEntry.modelClass);
-    }
-  } else {
-    relatedClass = relationEntry.modelClass;
-  }
-
+  const relatedClass = getModelClass(relationEntry.modelClass, processedClass.modelPaths);
   let result;
 
   //protect against endless recurstion
